refactor(...) : socket connect remote source pattern 변경
